requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.56.2
faiss-cpu==1.7.4
tiktoken==0.5.2
python-dotenv==1.0.0
pydantic==2.5.0